feat: add storage codegen
