+++ packages/cli/src/commands/codegen/codegen-portal.ts
@@ -1,0 +1,38 @@
+import { writeTemplate } from "@/commands/codegen/write-template";
+import { generateSchema } from "@gql.tada/cli-utils";
+import type { DotEnv } from "@settlemint/sdk-utils/validation";
+
+export async function codegenPortal(env: DotEnv) {
+  const gqlEndpoint = env.SETTLEMINT_PORTAL_GRAPHQL_ENDPOINT;
+  if (!gqlEndpoint) {
+    return;
+  }
+
+  const accessToken = env.SETTLEMINT_ACCESS_TOKEN;
+
+  await generateSchema({
+    input: gqlEndpoint,
+    output: "portal-schema.graphql",
+    tsconfig: undefined,
+    headers: {
+      "x-auth-token": accessToken,
+    },
+  });
+
+  const clientTemplate = `import { createServerPortalClient } from "@settlemint/sdk-portal";
+import type { introspection } from "../../portal-env.d.ts";
+
+export const { client: portalClient, graphql: portalGraphql } = createServerPortalClient<{
+  introspection: introspection;
+  disableMasking: true;
+  scalars: {
+    DateTime: Date;
+    JSON: Record<string, unknown>;
+  };
+}>({
+  instance: process.env.SETTLEMINT_PORTAL_GRAPHQL_ENDPOINT!,
+  accessToken: process.env.SETTLEMINT_ACCESS_TOKEN!,
+});`;
+
+  await writeTemplate(clientTemplate, "/lib/settlemint", "portal.ts");
+}
